django/requirements.txt
asgiref==3.8.1
coverage==7.6.0
Django==5.0.7
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.6
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
